move dummy code to add source folder
