+++ tests/unit_tests/test_dummy.py
@@ -6,7 +6,7 @@ class DummyTest:
 		assert 1==1
 
 	def test_add_numbers(self):
-		from dummy_code import add_numbers
+		from src.easydynamics.dummy_code import add_numbers
 		assert add_numbers(2, 3) == 5
 		assert add_numbers(-1, 1) == 0
 		assert add_numbers(0, 0) == 0
